+++ packages/cubejs-testing/src/birdbox.ts
@@ -48,7 +48,7 @@ interface Args {
   log: Log,
 }
 
-export type DriverType = 'postgresql' | 'postgres' | 'multidb' | 'materialize' | 'crate' | 'bigquery' | 'athena' | 'postgresql-cubestore' | 'firebolt' | 'questdb' | 'redshift' | 'databricks-jdbc' | 'prestodb' | 'mssql' | 'trino' | 'oracle' | 'ducksdb';
+export type DriverType = 'postgresql' | 'postgres' | 'multidb' | 'materialize' | 'crate' | 'bigquery' | 'athena' | 'postgresql-cubestore' | 'firebolt' | 'questdb' | 'redshift' | 'databricks-jdbc' | 'prestodb' | 'mssql' | 'trino' | 'oracle' | 'duckdb';
 
 export type Schemas = string[];
 
@@ -107,7 +107,7 @@ const driverNameToFolderNameMapper: Record<DriverType, string> = {
   mssql: 'mssql',
   trino: 'postgresql',
   oracle: 'oracle',
-  ducksdb: 'postgresql',
+  duckdb: 'postgresql',
 };
 
 /**